+++ src/globals.ts
@@ -15,9 +15,9 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { ISwaggerDocumentContext } from './contracts';
+import { ISwaggerDocumentContext } from "./contracts";
 
 /**
  * Store of Swagger documents.
  */
-export const swaggerDocuments: { [name: string]: ISwaggerDocumentContext } = {};
+export const swaggerDocuments: Record<string, ISwaggerDocumentContext> = {};